+++ lib/rubocop/cop/style/documentation.rb
@@ -7,6 +7,10 @@ module RuboCop
       # classes and modules. Classes with no body are exempt from the
       # check and so are namespace modules - modules that have nothing in
       # their bodies except classes or other other modules.
+      #
+      # The documentation requirement is annulled if the class or module has
+      # a "#:nodoc:" comment next to it. Likewise, "#:nodoc: all" does the
+      # same for all its children.
       class Documentation < Cop
         include AnnotationComment
 
@@ -38,6 +42,7 @@ module RuboCop
             next if node.type == :class && !body
             next if namespace?(body)
             next if associated_comment?(node, ast_with_comments)
+            next if nodoc?(node, ast_with_comments)
             add_offense(node, :keyword, format(MSG, node.type.to_s))
           end
         end
@@ -69,6 +74,26 @@ module RuboCop
           # As long as there's at least one comment line that isn't an
           # annotation, it's OK.
           ast_with_comments[node].any? { |comment| !annotation?(comment) }
+        end
+
+        # The :nodoc: comment is not actually associated with the class/module
+        # ifself but its first commentable child node. Unless the element is
+        # tagged with :nodoc:, the search proceeds to check its ancestors for
+        # :nodoc: all.
+        def nodoc?(node, ast_with_comments, require_all = false)
+          return false unless node
+          nodoc_node = node.children.last
+          return false unless nodoc_node
+
+          nodoc_node = nodoc_node.children.first while nodoc_node.type == :begin
+          comment = ast_with_comments[nodoc_node].first
+
+          if comment && comment.loc.line == node.loc.line
+            regex = /^#\s*:nodoc:#{"\s+all\s*$" if require_all}/
+            return true if comment.text =~ regex
+          end
+
+          nodoc?(node.ancestors.first, ast_with_comments, true)
         end
       end
     end